split off loan-handling code to its own file
